# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API des Commentaires",
        "version": "1.0"
    },
    "paths": {
...
        "/api/plans-nutritionnels": {
            "get": {
                "tags": [
                    "Plans Nutritionnels"
                ],
                "summary": "Lister tous les plans nutritionnels",
                "operationId": "890bedde826c7b3f94bc57636c67266a",
                "responses": {
                    "200": {
                        "description": "Liste des plans nutritionnels récupérée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PlanNutritionnel"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Plans Nutritionnels"
                ],
                "summary": "Créer un nouveau plan nutritionnel",
                "operationId": "eee863a80f0e24a7ae936d7efc05b0ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePlanNutritionnelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plan nutritionnel créé avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlanNutritionnel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/plans-nutritionnels/{id}": {
            "get": {
                "tags": [
                    "Plans Nutritionnels"
                ],
                "summary": "Afficher un plan spécifique",
                "operationId": "4a6e4f3a89cb8b4276f05462adb06011",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du plan récupérés avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlanNutritionnel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plan nutritionnel non trouvé."
                    }
                }
            },
            "put": {
                "tags": [
                    "Plans Nutritionnels"
                ],
                "summary": "Mettre à jour un plan nutritionnel",
                "operationId": "80faf2e6bbba620ed0b7b8f0e9db42f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePlanNutritionnelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plan nutritionnel mis à jour avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlanNutritionnel"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Plans Nutritionnels"
                ],
                "summary": "Supprimer un plan nutritionnel",
                "operationId": "1fb0d0feb3d95b629f8e8e6e681e7b32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan nutritionnel supprimé avec succès."
                    },
                    "404": {
                        "description": "Plan nutritionnel non trouvé."
                    }
                }
            }
        },
        "/api/programmes-entrainement": {
            "get": {
                "tags": [
                    "Programmes d'Entraînement"
                ],
                "summary": "Lister tous les programmes d'entraînement",
                "operationId": "9165d0f309ee3ef295a534dc49e0cff8",
                "responses": {
                    "200": {
                        "description": "Liste des programmes d'entraînement récupérée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProgrammeEntrainement"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Programmes d'Entraînement"
                ],
                "summary": "Créer un nouveau programme d'entraînement",
                "operationId": "be2f98b20681eff18bd9a0d182155bc8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProgrammeEntrainementRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Programme d'entraînement créé avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProgrammeEntrainement"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/programmes-entrainement/{id}": {
            "get": {
                "tags": [
                    "Programmes d'Entraînement"
                ],
                "summary": "Afficher un programme d'entraînement spécifique",
                "operationId": "2efb0998001abf54bfbb22ebac3c859b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du programme d'entraînement récupérés avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProgrammeEntrainement"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Programme d'entraînement non trouvé."
                    }
                }
            },
            "put": {
                "tags": [
                    "Programmes d'Entraînement"
                ],
                "summary": "Mettre à jour un programme d'entraînement",
                "operationId": "3a348083ba77b4c9cc1e3238559a92d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProgrammeEntrainementRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Programme d'entraînement mis à jour avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProgrammeEntrainement"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Programmes d'Entraînement"
                ],
                "summary": "Supprimer un programme d'entraînement",
                "operationId": "18624918f5900b7d3d5d0fb69020e569",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Programme d'entraînement supprimé avec succès."
                    },
                    "404": {
                        "description": "Programme d'entraînement non trouvé."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PlanNutritionnel": {
                "required": [
                    "id",
                    "name",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Plan de perte de poids"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un plan nutritionnel conçu pour aider à perdre du poids."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StorePlanNutritionnelRequest": {
                "required": [
                    "name",
                    "description"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Plan de prise de masse"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un plan nutritionnel conçu pour aider à prendre de la masse."
                    }
                },
                "type": "object"
            },
            "UpdatePlanNutritionnelRequest": {
                "required": [
                    "name",
                    "description"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Plan de perte de poids modifié"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un plan modifié pour aider à perdre du poids."
                    }
                },
                "type": "object"
            },
            "ProgrammeEntrainement": {
                "required": [
                    "id",
                    "title",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Programme de Force"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un programme d'entraînement axé sur le développement de la force."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreProgrammeEntrainementRequest": {
                "required": [
                    "title",
                    "description"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Programme d'Endurance"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un programme d'entraînement conçu pour améliorer l'endurance."
                    }
                },
                "type": "object"
            },
            "UpdateProgrammeEntrainementRequest": {
                "required": [
                    "title",
                    "description"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Programme de Force Modifié"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un programme d'entraînement modifié axé sur le développement de la force."
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}